# refactor(x/feegrant): collection migration (#16535) BACKPORT