--- a/x/feegrant/keeper/grpc_query_test.go
+++ b/x/feegrant/keeper/grpc_query_test.go
@@ -147,6 +147,25 @@ func (suite *KeeperTestSuite) TestFeeAllowances() {
 				suite.Require().Equal(resp.Allowances[0].Grantee, suite.addrs[1].String())
 			},
 		},
+		{
+			"valid query: expect multi grant",
+			&feegrant.QueryAllowancesRequest{
+				Grantee: suite.addrs[0].String(),
+			},
+			false,
+			func() {
+				suite.grantFeeAllowance(suite.addrs[1], suite.addrs[0])
+				suite.grantFeeAllowance(suite.addrs[2], suite.addrs[0])
+				suite.grantFeeAllowance(suite.addrs[3], suite.addrs[0])
+			},
+			func(resp *feegrant.QueryAllowancesResponse) {
+				suite.Require().Equal(len(resp.Allowances), 3)
+				for i, addr := range suite.addrs[1:4] {
+					resp.Allowances[i].Granter = suite.addrs[0].String()
+					resp.Allowances[i].Granter = addr.String()
+				}
+			},
+		},
 	}
 
 	for _, tc := range testCases {