--- a/x/feegrant/keeper/keeper.go
+++ b/x/feegrant/keeper/keeper.go
@@ -2,9 +2,11 @@ package keeper
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"time"
 
+	"cosmossdk.io/collections"
 	"cosmossdk.io/core/store"
 	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/log"
@@ -21,20 +23,39 @@ import (
 // Keeper manages state of all fee grants, as well as calculating approval.
 // It must have a codec with all available allowances registered.
 type Keeper struct {
-	cdc          codec.BinaryCodec
-	storeService store.KVStoreService
-	authKeeper   feegrant.AccountKeeper
-	bankKeeper   feegrant.BankKeeper
+	cdc               codec.BinaryCodec
+	storeService      store.KVStoreService
+	authKeeper        feegrant.AccountKeeper
+	bankKeeper        feegrant.BankKeeper
+	Schema            collections.Schema
+	FeeAllowance      collections.Map[collections.Pair[sdk.AccAddress, sdk.AccAddress], feegrant.Grant]
+	FeeAllowanceQueue collections.Map[collections.Triple[time.Time, sdk.AccAddress, sdk.AccAddress], bool]
 }
 
 var _ ante.FeegrantKeeper = &Keeper{}
 
 // NewKeeper creates a feegrant Keeper
 func NewKeeper(cdc codec.BinaryCodec, storeService store.KVStoreService, ak feegrant.AccountKeeper) Keeper {
+	sb := collections.NewSchemaBuilder(storeService)
+
 	return Keeper{
 		cdc:          cdc,
 		storeService: storeService,
 		authKeeper:   ak,
+		FeeAllowance: collections.NewMap(
+			sb,
+			feegrant.FeeAllowanceKeyPrefix,
+			"allowances",
+			collections.PairKeyCodec(sdk.LengthPrefixedAddressKey(sdk.AccAddressKey), sdk.LengthPrefixedAddressKey(sdk.AccAddressKey)), // nolint: staticcheck // sdk.LengthPrefixedAddressKey is needed to retain state compatibility
+			codec.CollValue[feegrant.Grant](cdc),
+		),
+		FeeAllowanceQueue: collections.NewMap(
+			sb,
+			feegrant.FeeAllowanceQueueKeyPrefix,
+			"allowances_queue",
+			collections.TripleKeyCodec(sdk.TimeKey, sdk.LengthPrefixedAddressKey(sdk.AccAddressKey), sdk.LengthPrefixedAddressKey(sdk.AccAddressKey)), // nolint: staticcheck // sdk.LengthPrefixedAddressKey is needed to retain state compatibility
+			collections.BoolValue,
+		),
 	}
 }
 
@@ -68,9 +89,6 @@ func (k Keeper) GrantAllowance(ctx context.Context, granter, grantee sdk.AccAddr
 		k.authKeeper.SetAccount(ctx, granteeAcc)
 	}
 
-	store := k.storeService.OpenKVStore(ctx)
-	key := feegrant.FeeAllowanceKey(granter, grantee)
-
 	exp, err := feeAllowance.ExpiresAt()
 	if err != nil {
 		return err
@@ -84,9 +102,7 @@ func (k Keeper) GrantAllowance(ctx context.Context, granter, grantee sdk.AccAddr
 
 	// if expiry is not nil, add the new key to pruning queue.
 	if exp != nil {
-		// `key` formed here with the prefix of `FeeAllowanceKeyPrefix` (which is `0x00`)
-		// remove the 1st byte and reuse the remaining key as it is
-		err = k.addToFeeAllowanceQueue(ctx, key[1:], exp)
+		err = k.FeeAllowanceQueue.Set(ctx, collections.Join3(*exp, grantee, granter), true)
 		if err != nil {
 			return err
 		}
@@ -97,13 +113,7 @@ func (k Keeper) GrantAllowance(ctx context.Context, granter, grantee sdk.AccAddr
 		return err
 	}
 
-	bz, err := k.cdc.Marshal(&grant)
-	if err != nil {
-		return err
-	}
-
-	err = store.Set(key, bz)
-	if err != nil {
+	if err := k.FeeAllowance.Set(ctx, collections.Join(grantee, granter), grant); err != nil {
 		return err
 	}
 
@@ -120,9 +130,6 @@ func (k Keeper) GrantAllowance(ctx context.Context, granter, grantee sdk.AccAddr
 
 // UpdateAllowance updates the existing grant.
 func (k Keeper) UpdateAllowance(ctx context.Context, granter, grantee sdk.AccAddress, feeAllowance feegrant.FeeAllowanceI) error {
-	store := k.storeService.OpenKVStore(ctx)
-	key := feegrant.FeeAllowanceKey(granter, grantee)
-
 	_, err := k.getGrant(ctx, granter, grantee)
 	if err != nil {
 		return err
@@ -133,13 +140,7 @@ func (k Keeper) UpdateAllowance(ctx context.Context, granter, grantee sdk.AccAdd
 		return err
 	}
 
-	bz, err := k.cdc.Marshal(&grant)
-	if err != nil {
-		return err
-	}
-
-	err = store.Set(key, bz)
-	if err != nil {
+	if err := k.FeeAllowance.Set(ctx, collections.Join(grantee, granter), grant); err != nil {
 		return err
 	}
 
@@ -161,9 +162,7 @@ func (k Keeper) revokeAllowance(ctx context.Context, granter, grantee sdk.AccAdd
 		return err
 	}
 
-	store := k.storeService.OpenKVStore(ctx)
-	key := feegrant.FeeAllowanceKey(granter, grantee)
-	err = store.Delete(key)
+	err = k.FeeAllowance.Remove(ctx, collections.Join(grantee, granter))
 	if err != nil {
 		return err
 	}
@@ -174,7 +173,7 @@ func (k Keeper) revokeAllowance(ctx context.Context, granter, grantee sdk.AccAdd
 	}
 
 	if exp != nil {
-		if err := store.Delete(feegrant.FeeAllowancePrefixQueue(exp, feegrant.FeeAllowanceKey(grantee, granter)[1:])); err != nil {
+		if err := k.FeeAllowanceQueue.Remove(ctx, collections.Join3(*exp, grantee, granter)); err != nil {
 			return err
 		}
 	}
@@ -193,7 +192,7 @@ func (k Keeper) revokeAllowance(ctx context.Context, granter, grantee sdk.AccAdd
 // If there is none, it returns nil, nil.
 // Returns an error on parsing issues
 func (k Keeper) GetAllowance(ctx context.Context, granter, grantee sdk.AccAddress) (feegrant.FeeAllowanceI, error) {
-	grant, err := k.getGrant(ctx, granter, grantee)
+	grant, err := k.FeeAllowance.Get(ctx, collections.Join(grantee, granter))
 	if err != nil {
 		return nil, err
 	}
@@ -203,19 +202,8 @@ func (k Keeper) GetAllowance(ctx context.Context, granter, grantee sdk.AccAddres
 
 // getGrant returns entire grant between both accounts
 func (k Keeper) getGrant(ctx context.Context, granter, grantee sdk.AccAddress) (*feegrant.Grant, error) {
-	store := k.storeService.OpenKVStore(ctx)
-	key := feegrant.FeeAllowanceKey(granter, grantee)
-	bz, err := store.Get(key)
+	feegrant, err := k.FeeAllowance.Get(ctx, collections.Join(grantee, granter))
 	if err != nil {
-		return nil, err
-	}
-
-	if len(bz) == 0 {
-		return nil, sdkerrors.ErrNotFound.Wrap("fee-grant not found")
-	}
-
-	var feegrant feegrant.Grant
-	if err := k.cdc.Unmarshal(bz, &feegrant); err != nil {
 		return nil, err
 	}
 
@@ -230,14 +218,12 @@ func (k Keeper) IterateAllFeeAllowances(ctx context.Context, cb func(grant feegr
 	iter := storetypes.KVStorePrefixIterator(runtime.KVStoreAdapter(store), feegrant.FeeAllowanceKeyPrefix)
 	defer iter.Close()
 
-	stop := false
-	for ; iter.Valid() && !stop; iter.Next() {
-		bz := iter.Value()
-		var feeGrant feegrant.Grant
-		if err := k.cdc.Unmarshal(bz, &feeGrant); err != nil {
-			return err
-		}
-		stop = cb(feeGrant)
+	err := k.FeeAllowance.Walk(ctx, nil, func(key collections.Pair[sdk.AccAddress, sdk.AccAddress], grant feegrant.Grant) (stop bool, err error) {
+		return cb(grant), nil
+	})
+
+	if err != nil && !errors.Is(err, collections.ErrInvalidIterator) {
+		return err
 	}
 
 	return nil
@@ -323,39 +309,28 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*feegrant.GenesisState, erro
 	}, err
 }
 
-func (k Keeper) addToFeeAllowanceQueue(ctx context.Context, grantKey []byte, exp *time.Time) error {
-	store := k.storeService.OpenKVStore(ctx)
-	return store.Set(feegrant.FeeAllowancePrefixQueue(exp, grantKey), []byte{})
-}
-
 // RemoveExpiredAllowances iterates grantsByExpiryQueue and deletes the expired grants.
 func (k Keeper) RemoveExpiredAllowances(ctx context.Context, limit int32) error {
 	exp := sdk.UnwrapSDKContext(ctx).BlockTime()
-	store := k.storeService.OpenKVStore(ctx)
-	iterator, err := store.Iterator(feegrant.FeeAllowanceQueueKeyPrefix, storetypes.InclusiveEndBytes(feegrant.AllowanceByExpTimeKey(&exp)))
-	var count int32
-	if err != nil {
+	rng := collections.NewPrefixUntilTripleRange[time.Time, sdk.AccAddress, sdk.AccAddress](exp)
+
+	err := k.FeeAllowanceQueue.Walk(ctx, rng, func(key collections.Triple[time.Time, sdk.AccAddress, sdk.AccAddress], value bool) (stop bool, err error) {
+		grantee, granter := key.K2(), key.K3()
+
+		if err := k.FeeAllowance.Remove(ctx, collections.Join(grantee, granter)); err != nil {
+			return true, err
+		}
+
+		if err := k.FeeAllowanceQueue.Remove(ctx, key); err != nil {
+			return true, err
+		}
+
+		return false, nil
+	})
+
+	if err != nil && !errors.Is(err, collections.ErrInvalidIterator) {
 		return err
 	}
-	defer iterator.Close()
 
-	for ; iterator.Valid(); iterator.Next() {
-		err = store.Delete(iterator.Key())
-		if err != nil {
-			return err
-		}
-
-		granter, grantee := feegrant.ParseAddressesFromFeeAllowanceQueueKey(iterator.Key())
-		err = store.Delete(feegrant.FeeAllowanceKey(granter, grantee))
-		if err != nil {
-			return err
-		}
-
-		// limit the amount of iterations to avoid taking too much time
-		count++
-		if count == limit {
-			return nil
-		}
-	}
 	return nil
 }