--- a/x/feegrant/keeper/msg_server_test.go
+++ b/x/feegrant/keeper/msg_server_test.go
@@ -227,7 +227,7 @@ func (suite *KeeperTestSuite) TestRevokeAllowance() {
 			},
 			func() {},
 			true,
-			"fee-grant not found",
+			"not found",
 		},
 		{
 			"success: revoke fee allowance",
@@ -269,7 +269,7 @@ func (suite *KeeperTestSuite) TestRevokeAllowance() {
 			},
 			func() {},
 			true,
-			"fee-grant not found",
+			"not found",
 		},
 	}