--- a/x/feegrant/module/abci_test.go
+++ b/x/feegrant/module/abci_test.go
@@ -84,7 +84,7 @@ func TestFeegrantPruning(t *testing.T) {
 	})
 	require.NoError(t, err)
 	require.NotNil(t, res)
-	require.Len(t, res.Allowances, 3)
+	require.Len(t, res.Allowances, 2)
 
 	testCtx.Ctx = testCtx.Ctx.WithBlockTime(now.AddDate(0, 0, 2))
 	require.NoError(t, module.EndBlocker(testCtx.Ctx, feeGrantKeeper))