@technicallyty 2025-04-11
can you add this test case? 


x/feegrant/keeper/grpc_query_test.go:150
```go
		{
			"valid query: expect multi grant",
			&feegrant.QueryAllowancesRequest{
				Grantee: suite.addrs[0].String(),
			},
			false,
			func() {
				suite.grantFeeAllowance(suite.addrs[1], suite.addrs[0])
				suite.grantFeeAllowance(suite.addrs[2], suite.addrs[0])
				suite.grantFeeAllowance(suite.addrs[3], suite.addrs[0])
			},
			func(resp *feegrant.QueryAllowancesResponse) {
				suite.Require().Equal(len(resp.Allowances), 3)
				for i, addr := range suite.addrs[1:4] {
					resp.Allowances[i].Granter = suite.addrs[0].String()
					resp.Allowances[i].Granter = addr.String()
				}
			},
		},
```