@DuongNguyen62025 2025-04-15
I tried running the benchmark with the following code:
```go
func BenchmarkGetAllowance(b *testing.B) {
	b.ReportAllocs()
	var keeper keeper.Keeper
	app, err := simtestutil.Setup(
		depinject.Configs(
			depinject.Supply(log.NewNopLogger()),
			appConfig,
		),
		&keeper,
	)
	require.NoError(b, err)

	ctx := app.NewContext(false)
	granter := sdk.AccAddress([]byte("granter________________"))
	grantee := sdk.AccAddress([]byte("grantee________________"))

	now := ctx.BlockTime()
	if err := keeper.GrantAllowance(ctx, granter, grantee, &feegrant.BasicAllowance{
		SpendLimit: sdk.NewCoins(sdk.NewInt64Coin("atom", 555)),
		Expiration: &now,
	}); err != nil {
		b.Fatal(err)
	}

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		_, err := keeper.GetAllowance(ctx, granter, grantee)
		if err != nil {
			b.Fatal(err)
		}
	}
}

```
results:

`main`: 
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkGetAllowance$ cosmossdk.io/x/feegrant/keeper

goos: darwin
goarch: arm64
pkg: cosmossdk.io/x/feegrant/keeper
cpu: Apple M2 Pro
BenchmarkGetAllowance-10    	  705301	      1730 ns/op	    2112 B/op	      35 allocs/op
PASS
ok  	cosmossdk.io/x/feegrant/keeper	1.908s
```
`this PR`:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkGetAllowance$ cosmossdk.io/x/feegrant/keeper

goos: darwin
goarch: arm64
pkg: cosmossdk.io/x/feegrant/keeper
cpu: Apple M2 Pro
BenchmarkGetAllowance-10    	  647562	      1715 ns/op	    2080 B/op	      32 allocs/op
PASS
ok  	cosmossdk.io/x/feegrant/keeper	1.822s
```

**1. Processing time (ns/op)**
`main`: 1730 ns/op
`this PR`: 1715 ns/op
Processing time is slightly reduced by 15 ns/op (about 0.87% improvement). Although the improvement is not very large in terms of average time per operation, it is still a sign that optimizations or improvements in fee allowance access have a positive effect.

**2. Memory Usage (B/op)**
`main`: 2112 B/op
`this PR`: 2080 B/op
The 32 B/op reduction (about 1.5% reduction) shows that the new version consumes less memory than the old version. This saving may come from the improved data management in the feegrant module when switching to the new collections system.

**3. Allocations (allocs/op)**
`main`: 35 allocs/op
`this PR`: 32 allocs/op
The new version reduces allocs/op by 3 (about 8.6% reduction), which indicates that the new implementation creates fewer objects during the GetAllowance operation. Reducing allocations often results in less heap pressure and less frequency of garbage collector runs, which can improve overall performance under high load scenarios.